add workaround in getting SMART
From WD SN550, controller's SMART / Health information cannot be retrieved by Get Log Page with NSID = FFFFFFFFh.
So, only for WD SN550, the NSID to get it is changed to 0 (zero).
Model Name and Firmware Revision in Identify Controller data are used to identify the target device to enable this workaround.